make cache for offline

+++ app/src/main/java/com/learntodroid/simplealarmclock/data/DefaultAlarmRepositoryImpl.java
@@ -1,9 +1,10 @@
 package com.learntodroid.simplealarmclock.data;
 
 import android.app.Application;
-import android.util.Log;
+import android.content.BroadcastReceiver;
+import android.content.Context;
 
-import androidx.multidex.BuildConfig;
+import com.learntodroid.simplealarmclock.broadcastreceiver.NetworkChangeReceiver;
 
 import java.util.List;
 
@@ -15,19 +16,23 @@ import io.reactivex.rxjava3.functions.Function;
 public class DefaultAlarmRepositoryImpl implements AlarmRepository {
     private final AlarmNetwork network;
     private final AlarmLocal local;
+    private final CacheAlarmLocal cacheLocal;
+    private final Context context;
 
     private static DefaultAlarmRepositoryImpl instance = null;
-    public static DefaultAlarmRepositoryImpl getInstance(Application application){
-        if(instance == null) {
-            instance = new DefaultAlarmRepositoryImpl(application);
-        }
-        return  instance;
-    }
 
     private DefaultAlarmRepositoryImpl(Application application) {
-        Log.e("cccccccccccccc", BuildConfig.FLAVOR);
         network = NetworkDI.getAlarmNetwork();
         local = AlarmDatabase.getDatabase(application).alarmDao();
+        cacheLocal = AlarmDatabase.getDatabase(application).cacheAlarmDao();
+        context = application.getBaseContext();
+    }
+
+    public static DefaultAlarmRepositoryImpl getInstance(Application application) {
+        if (instance == null) {
+            instance = new DefaultAlarmRepositoryImpl(application);
+        }
+        return instance;
     }
 
     @Override
@@ -38,30 +43,52 @@ public class DefaultAlarmRepositoryImpl implements AlarmRepository {
     @Override
     public void onClearListener() {
         network.onClearListener();
+        cacheLocal.deleteAll();
     }
 
     @Override
     public Single<String> insert(Alarm alarm) {
-        return network.insert(alarm);
+        if (NetworkChangeReceiver.isOnline(context)) {
+            return network.insert(alarm);
+        }
+        return cacheLocal.insert(alarm).map(s -> {
+            local.insert(alarm);
+            return s;
+        });
     }
 
     @Override
     public Single<String> update(Alarm alarm) {
-        return network.update(alarm);
+        if (NetworkChangeReceiver.isOnline(context)) {
+            return network.update(alarm);
+        }
+        return cacheLocal.update(alarm).map(s -> {
+            local.update(alarm);
+            return s;
+        });
     }
 
     @Override
     public Single<String> delete(Alarm alarm) {
-        return network.delete(alarm);
+        if (NetworkChangeReceiver.isOnline(context)) {
+            return network.delete(alarm);
+        }
+        return cacheLocal.delete(alarm).map(s -> {
+            local.delete(alarm);
+            return s;
+        });
     }
 
     @Override
     public Observable<List<Alarm>> getAlarms() {
-        return network.getAlarms().flatMap((Function<List<Alarm>, ObservableSource<List<Alarm>>>) alarms -> {
-            local.deleteAll();
-            local.insert(alarms.toArray(new Alarm[0]));
-            return local.getAlarms();
-        });
+        if (NetworkChangeReceiver.isOnline(context)) {
+            return network.getAlarms().flatMap((Function<List<Alarm>, ObservableSource<List<Alarm>>>) alarms -> {
+                local.deleteAll();
+                local.insert(alarms.toArray(new Alarm[0]));
+                return local.getAlarms();
+            });
+        }
+        return local.getAlarms();
     }
 
     @Override
@@ -70,7 +97,10 @@ public class DefaultAlarmRepositoryImpl implements AlarmRepository {
     }
 
     @Override
-    public void refresh() {
-        network.refresh();
+    public void refresh() throws NetworkChangeReceiver.NoConnectInternet {
+        if (NetworkChangeReceiver.isOnline(context)) {
+            network.refresh();
+        }
+        throw new NetworkChangeReceiver.NoConnectInternet();
     }
 }
